Added initial implement of the audacious plugin for audacious 1.4.x or higher. There are still a few things to do such as add a configuration GUI, but these things are marked via TODO in the source files.
I also added the autoconf/automake building system.  To generate everything
you need, run ./bootstrap in the top level directory, it will generate the
configure script.  Then run ./configure.  It will generate Makefile.unix, so
build it via make -f Makefile.unix.  Everything should be setup properly, after
building, execute 'make -f Makefile.unix install' to install.




git-svn-id: https://vgmstream.svn.sourceforge.net/svnroot/vgmstream@165 51a99a44-fe44-0410-b1ba-c3e57ba2b86b

+++ src/vgmstream.c
@@ -526,12 +526,12 @@ void describe_vgmstream(VGMSTREAM * vgmstream, char * desc, int length) {
 
     if (vgmstream->layout_type == layout_interleave || vgmstream->layout_type == layout_interleave_shortblock) {
         snprintf(temp,TEMPSIZE,"interleave: %#x bytes\n",
-                vgmstream->interleave_block_size);
+                (int32_t)vgmstream->interleave_block_size);
         concatn(length,desc,temp);
 
         if (vgmstream->layout_type == layout_interleave_shortblock) {
             snprintf(temp,TEMPSIZE,"last block interleave: %#x bytes\n",
-                    vgmstream->interleave_smallblock_size);
+                    (int32_t)vgmstream->interleave_smallblock_size);
             concatn(length,desc,temp);
         }
     }
